# serviceos-contracts/src/main/resources/openapi/serviceos-core-v1.yaml
openapi: 3.1.0
info:
  title: ServiceOS Core API
  version: 0.4.0
  description: 首条工程纵向切片的机器可读契约；实现与客户端均不得绕过本契约自行猜测字段。
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
  /projects:
    post:
      operationId: createProject
      summary: 创建运营项目
      description: 同一租户、操作和 Idempotency-Key 下，相同请求返回同一资源；不同请求返回冲突。
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: 项目创建成功或幂等重放成功
          headers:
            Location:
              schema: { type: string }
            X-Correlation-Id:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
  /tasks/{taskId}:claim:
    post:
      operationId: claimHumanTask
      summary: 领取 READY 人工任务
      description: 主体来自 JWT；相同幂等键和请求返回首次冻结响应，陈旧版本或并发领取返回 409。
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200': { $ref: '#/components/responses/HumanTaskCommandReceipt' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /tasks/{taskId}:assign-candidates:
    post:
      operationId: assignTaskCandidates
      summary: 为 READY 人工任务冻结 USER 候选快照
      description: 替换旧候选快照并递增 Task 版本；角色/组织策略必须先解析为稳定 USER ID。
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AssignTaskCandidatesRequest' }
      responses:
        '200':
          description: 候选快照已冻结或返回首次幂等响应
          headers:
            ETag: { required: true, schema: { type: string } }
            X-Correlation-Id: { required: true, schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskAssignmentBatchReceipt' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /tasks/{taskId}:start:
    post:
      operationId: startHumanTask
      summary: 当前领取人启动人工任务
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200': { $ref: '#/components/responses/HumanTaskCommandReceipt' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /tasks/{taskId}:complete:
    post:
      operationId: completeHumanTask
      summary: 当前执行人完成人工工作流任务
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CompleteHumanTaskRequest' }
      responses:
        '200': { $ref: '#/components/responses/HumanTaskCommandReceipt' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /tasks/{taskId}:release:
    post:
      operationId: releaseHumanTask
      summary: 当前责任人在 start 前释放已领取任务
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReleaseHumanTaskRequest' }
      responses:
        '200': { $ref: '#/components/responses/ReleasedHumanTaskReceipt' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /files/upload-sessions:
    post:
      operationId: beginFileUpload
      summary: 创建受限文件上传会话
      description: >-
        服务端生成私有 object key，返回绑定精确大小、声明 MIME 和有效期的短期上传授权；
        相同 Idempotency-Key 与请求会返回同一 uploadSessionId。
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BeginUploadRequest'
      responses:
        '201':
          description: 上传会话创建或幂等重放成功
          headers:
            Location:
              schema: { type: string }
            X-Correlation-Id:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadSession'
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /files/upload-sessions/{uploadSessionId}:finalize:
    post:
      operationId: finalizeFileUpload
      summary: 校验对象并创建不可变文件
      description: >-
        事务外重新读取对象的真实大小、SHA-256 与魔数 MIME；通过后创建 QUARANTINED/PENDING_SCAN
        文件并在同一事务调度扫描任务。对象存储回调不能调用此能力。
      parameters:
        - name: uploadSessionId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FinalizeUploadRequest'
      responses:
        '201':
          description: 文件已创建或 finalize 幂等重放成功；扫描通过前仍不可下载
          headers:
            Location:
              schema: { type: string }
            X-Correlation-Id:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoredFile'
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '410': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
  /files/{fileId}/download-authorizations:
    post:
      operationId: authorizeFileDownload
      summary: 申请文件短期下载授权
      description: 每次实时校验 RoleGrant；仅 AVAILABLE 文件可获得短期私有下载 URL，目的写入审计。
      parameters:
        - name: fileId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorizeDownloadRequest'
      responses:
        '201':
          description: 短期下载授权已签发
          headers:
            Location:
              schema: { type: string }
            X-Correlation-Id:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DownloadAuthorization'
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '423': { $ref: '#/components/responses/Problem' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 由受信 OIDC 提供方签发，必须包含 subject、tenant_id；token 中的能力声明不能替代 ServiceOS RoleGrant。
  parameters:
    TaskId:
      name: taskId
      in: path
      required: true
      schema: { type: string, format: uuid }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 160
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      schema:
        type: string
        maxLength: 128
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: 双引号包裹的正整数 Task 聚合版本，例如 `"3"`。
      schema:
        type: string
        pattern: '^"[1-9][0-9]*"$'
  schemas:
    CreateProjectRequest:
      type: object
      additionalProperties: false
      required: [code, clientId, name, startsOn]
      properties:
        code:
          type: string
          pattern: '^[A-Za-z0-9][A-Za-z0-9_-]*$'
          maxLength: 64
        clientId:
          type: string
          minLength: 1
          maxLength: 128
        name:
          type: string
          minLength: 1
          maxLength: 200
        startsOn:
          type: string
          format: date
        endsOn:
          type: [string, 'null']
          format: date
    Project:
      type: object
      additionalProperties: false
      required: [id, tenantId, code, clientId, name, startsOn, status, version, createdAt]
      properties:
        id: { type: string, format: uuid }
        tenantId: { type: string }
        code: { type: string }
        clientId: { type: string }
        name: { type: string }
        startsOn: { type: string, format: date }
        endsOn: { type: [string, 'null'], format: date }
        status:
          type: string
          enum: [DRAFT, ACTIVE, SUSPENDED, CLOSED]
        version: { type: integer, format: int64, minimum: 1 }
        createdAt: { type: string, format: date-time }
    CompleteHumanTaskRequest:
      type: object
      additionalProperties: false
      required: [resultRef, resultDigest]
      properties:
        resultRef: { type: string, minLength: 1, maxLength: 500 }
        resultDigest: { type: string, pattern: '^[0-9a-f]{64}$' }
    AssignTaskCandidatesRequest:
      type: object
      additionalProperties: false
      required: [candidatePrincipalIds, sourceType, sourceId]
      properties:
        candidatePrincipalIds:
          type: array
          minItems: 1
          maxItems: 100
          items: { type: string, minLength: 1, maxLength: 128 }
        sourceType: { enum: [ASSIGNEE_POLICY, MANUAL] }
        sourceId: { type: string, minLength: 1, maxLength: 160 }
    ReleaseHumanTaskRequest:
      type: object
      additionalProperties: false
      required: [reasonCode]
      properties:
        reasonCode: { type: string, pattern: '^[A-Z][A-Z0-9_]{1,99}$' }
    TaskAssignmentBatchReceipt:
      type: object
      additionalProperties: false
      required: [assignmentBatchId, taskId, candidateCount, taskVersion, assignedAt]
      properties:
        assignmentBatchId: { type: string, format: uuid }
        taskId: { type: string, format: uuid }
        candidateCount: { type: integer, minimum: 1, maximum: 100 }
        taskVersion: { type: integer, format: int64, minimum: 2 }
        assignedAt: { type: string, format: date-time }
    HumanTaskCommandReceipt:
      type: object
      additionalProperties: false
      required: [taskId, status, actorId, version, occurredAt]
      properties:
        taskId: { type: string, format: uuid }
        status: { enum: [CLAIMED, RUNNING, COMPLETED] }
        actorId: { type: string, minLength: 1, maxLength: 128 }
        version: { type: integer, format: int64, minimum: 2 }
        occurredAt: { type: string, format: date-time }
    ReleasedHumanTaskReceipt:
      type: object
      additionalProperties: false
      required: [taskId, status, actorId, version, occurredAt]
      properties:
        taskId: { type: string, format: uuid }
        status: { const: READY }
        actorId: { type: string, minLength: 1, maxLength: 128 }
        version: { type: integer, format: int64, minimum: 2 }
        occurredAt: { type: string, format: date-time }
    BeginUploadRequest:
      type: object
      additionalProperties: false
      required:
        [businessContextType, businessContextId, originalFileName, declaredMimeType, expectedSize, expectedSha256]
      properties:
        businessContextType:
          type: string
          pattern: '^[A-Za-z][A-Za-z0-9_.-]*$'
          maxLength: 80
        businessContextId: { type: string, minLength: 1, maxLength: 160 }
        originalFileName: { type: string, minLength: 1, maxLength: 255 }
        declaredMimeType: { type: string, minLength: 3, maxLength: 255 }
        expectedSize: { type: integer, format: int64, minimum: 1 }
        expectedSha256: { type: string, pattern: '^[0-9A-Fa-f]{64}$' }
    FinalizeUploadRequest:
      type: object
      additionalProperties: false
      required: [actualSha256, finalizeCommandId]
      properties:
        actualSha256: { type: string, pattern: '^[0-9A-Fa-f]{64}$' }
        finalizeCommandId: { type: string, minLength: 1, maxLength: 160 }
    AuthorizeDownloadRequest:
      type: object
      additionalProperties: false
      required: [purpose]
      properties:
        purpose: { type: string, minLength: 1, maxLength: 500 }
    UploadSession:
      type: object
      additionalProperties: false
      required:
        [uploadSessionId, fileId, status, uploadMethod, uploadUrl, requiredHeaders,
         uploadAuthorizationExpiresAt, sessionExpiresAt]
      properties:
        uploadSessionId: { type: string, format: uuid }
        fileId: { type: string, format: uuid }
        status:
          enum: [CREATED, UPLOADING, FINALIZING, COMPLETED, EXPIRED, FAILED]
        uploadMethod: { type: [string, 'null'], enum: [PUT, null] }
        uploadUrl: { type: [string, 'null'], format: uri }
        requiredHeaders:
          type: object
          additionalProperties: { type: string }
        uploadAuthorizationExpiresAt: { type: [string, 'null'], format: date-time }
        sessionExpiresAt: { type: string, format: date-time }
    StoredFile:
      type: object
      additionalProperties: false
      required:
        [fileId, tenantId, originalFileName, checksumSha256, size, declaredMimeType,
         detectedMimeType, lifecycleStatus, quarantineReason, createdAt, version]
      properties:
        fileId: { type: string, format: uuid }
        tenantId: { type: string, minLength: 1, maxLength: 64 }
        originalFileName: { type: string, minLength: 1, maxLength: 255 }
        checksumSha256: { type: string, pattern: '^[0-9a-f]{64}$' }
        size: { type: integer, format: int64, minimum: 1 }
        declaredMimeType: { type: string }
        detectedMimeType: { type: string }
        lifecycleStatus: { enum: [QUARANTINED, AVAILABLE, INVALIDATED] }
        quarantineReason: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
        version: { type: integer, format: int64, minimum: 1 }
    DownloadAuthorization:
      type: object
      additionalProperties: false
      required: [authorizationId, fileId, method, downloadUrl, requiredHeaders, expiresAt]
      properties:
        authorizationId: { type: string, format: uuid }
        fileId: { type: string, format: uuid }
        method: { type: string, const: GET }
        downloadUrl: { type: string, format: uri }
        requiredHeaders:
          type: object
          additionalProperties: { type: string }
        expiresAt: { type: string, format: date-time }
    Problem:
      type: object
      required: [type, title, status, errorCode]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        errorCode: { type: string }
        correlationId: { type: [string, 'null'] }
  responses:
    HumanTaskCommandReceipt:
      description: 命令成功或幂等重放成功
      headers:
        ETag:
          required: true
          schema: { type: string }
        X-Correlation-Id:
          required: true
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/HumanTaskCommandReceipt' }
    ReleasedHumanTaskReceipt:
      description: 释放成功或幂等重放成功
      headers:
        ETag:
          required: true
          schema: { type: string }
        X-Correlation-Id:
          required: true
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ReleasedHumanTaskReceipt' }
    Problem:
      description: RFC 9457 Problem Details
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'